docs/agent_ai/tutorials/first_agent/requirements.txt
cohere>=5.0.0
python-dotenv>=1.0.1
requests>=2.31.0
pydantic>=2.7.0
rich>=13.7.1
pytest>=8.2.0